apply strategy pattern
