+++ .gitignore
@@ -39,6 +39,7 @@ build-iPhoneSimulator/
 ## Environment normalization:
 /.bundle/
 /vendor/bundle
+/vendor/bundler
 /lib/bundler/man/
 
 # for a library or gem, you might want to ignore these files since the code is
@@ -50,3 +51,5 @@ build-iPhoneSimulator/
 # unless supporting rvm < 1.11.0 or doing something fancy, ignore this:
 .rvmrc
 
+# intelli J
+.idea
