+++ model/strategy/formatter/confirm_message.rb
@@ -1,0 +1,25 @@
+
+class ConfirmMessage
+  def output_message(context)
+    {
+        "type": "template",
+        "altText": "this is a confirm template",
+        "template": {
+            "type": "confirm",
+            "text": "Are you sure?",
+            "actions": [
+                {
+                    "type": "message",
+                    "label": "Yes",
+                    "text": "yes"
+                },
+                {
+                    "type": "message",
+                    "label": "No",
+                    "text": "no"
+                }
+            ]
+        }
+    }
+  end
+end
