+++ model/strategy/formatter/default_message.rb
@@ -1,0 +1,9 @@
+
+class DefaultMessage
+  def output_message(context)
+    {
+        type: "text",
+        text: context.event.message['text']
+    }
+  end
+end
