+++ model/strategy/formatter/meeting_message.rb
@@ -1,0 +1,9 @@
+
+class MeetingMessage
+  def output_message(context)
+    {
+        type: 'text',
+        text: "次回ミーティング日程\n日時：11月1日（火）18:00\n場所：３階リフレッシュルーム\n内容：開発段階のプレビューと追加機能などに関する話し合い"
+    }
+  end
+end
