+++ model/strategy/formatter/sticker_message.rb
@@ -1,0 +1,9 @@
+class StickerMessage
+  def output_message(context)
+    {
+        type: "sticker",
+        packageId: 1,
+        stickerId: 1
+    }
+  end
+end
